+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to [Bach](https://github.com/sormuras/bach) will be document
 The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [JEP 223: New Version-String Scheme](https://openjdk.java.net/jeps/223).
 
-## Version [15-ea] unreleased
+## Version [Unreleased]
 
 Nothing notable, yet.
 
-[15-ea]: https://github.com/sormuras/bach/tree/main
+## Version [15-ea+1]
+
+Initial pre-release of Bach 15.
+
+### Added
+
+- Module `com.github.sormuras.bach` with generic tool-related classes in package `com.github.sormuras.bach`.
+
+[Unreleased]: https://github.com/sormuras/bach/compare/15-ea+1...HEAD
+[15-ea+1]: https://github.com/sormuras/bach/releases/tag/15-ea+1