+++ CHANGELOG.md
@@ -8,13 +8,14 @@ and this project adheres to [JEP 223: New Version-String Scheme](https://openjdk
 
 Nothing notable, yet.
 
-## Version [15-ea+1]
+## Version [15-ea+1] - 2020-10-12
 
 Initial pre-release of Bach 15.
 
 ### Added
 
-- Module `com.github.sormuras.bach` with generic tool-related classes in package `com.github.sormuras.bach`.
+- Module `com.github.sormuras.bach` exporting its only package with the same name
+- Package `com.github.sormuras.bach` with generic tool-related classes
 
 [Unreleased]: https://github.com/sormuras/bach/compare/15-ea+1...HEAD
 [15-ea+1]: https://github.com/sormuras/bach/releases/tag/15-ea+1