+++ CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 
 This project adheres to [JEP 223: New Version-String Scheme](https://openjdk.java.net/jeps/223).
 
-## Version [Early Access]
+## Version [16-ea]
 
 Nothing notable, yet.
+
+## Version [16.0.1] released 2021-01-03
+
+Minor build and output fixes.
 
 ## Version [16] released 2020-12-27
 
 Initial release.
 
-[Early Access]: https://github.com/sormuras/bach/releases/tag/early-access
-[16]: https://github.com/sormuras/bach/compare/16...main
+[16-ea]: https://github.com/sormuras/bach/releases/tag/16-ea
+[16.0.1]: https://github.com/sormuras/bach/compare/16.0.1...main
+[16]: https://github.com/sormuras/bach/compare/16...16.0.1